Add the error messages (that don't work) and the Hub functionality.
The error message function doesn't work because every time you log out it rebuilds the page twice because it builds it then detects an auth state change nad builds it again. A fix to this would be to only log out once you click the log in button. For example, when you log in with a non-Milton.edu email, it logs you out right away and sends you to the login screen with an error message but it could instead send you to the page with the error message and log you out once you clicked log in.